finished ground types and working on higher-kinded types
